DO NOT USE IN YOUR ROM!!! Read Commit Message
DO NOT MERGE in your own ROM! This isn't because we want to be assholes
about it but because OWM allows only 60 free calls per minute and the
more people that use this API key, the more calls are send their way. We
end up losing access and if you use this then you will lose access.

Nobody wins if you decide to use our keys.

Create your own, takes 30 seconds
https://openweathermap.org/appid#get

f6467bf